Forecast/FinCast-fts/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
beartype==0.20.2
Bottleneck==1.4.2
Brotli==1.1.0
certifi==2025.1.31
click==8.1.8
einops==0.8.1
einx==0.3.0
huggingface-hub==0.30.1
jax==0.5.3
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
wandb==0.19.9
yfinance==0.2.55
utilsforecast==0.2.12
CoLT5-attention==0.11.1
tensorflow==2.19.0
refinitiv-data==1.6.2
peft
torch